requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
Django==2.1
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==3.0.0
django-summernote==0.8.11.6
importlib-metadata==2.1.1
lorem==0.1.1
Pillow==8.2.0
pkg-resources==0.0.0
pytz==2021.1
soupsieve==2.2.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
zipp==3.4.1
